# [Unity][DLight] Introduce Specific Rule for RMSNorm