@Celve 2024-01-06
> Does the result in Old refer to dlight or cutlass?

It refers to dlight.