@vinx13 2024-01-06
thanks for the clarification, do we have numbers comparing with cutlass?